Video Demo
we exited the game first by pressing zero. 
Then we re-executed the program by entering "yes".
Then we played for easy level and lost.
Upon finishing the game we chose to continue playing.
We played for level 2 and lost.
Upon finishing the game we chose to continue playing.
We played for level 3 and lost.
Upon finishing the game we chose to exit.